tirpc: Fix numerous errors and warnings.
Apart from minor printf() formatting fixes and removal of unused
variables this commit adds "#include <pthread.h>" to testcases that are
utilizing pthreads and fixes argument passing to the pthread start
function.

There should be no changes in functionality except for one testcase that
was missing "return test_status;" at the end of the main().

+++ testcases/network/rpc/rpc-tirpc/tests_pack/rpc_suite/rpc/rpc_addrmanagmt_pmap_rmtcall/rpc_pmap_rmtcall.c
@@ -72,7 +72,7 @@ int main(int argn, char *argc[])
 	tv.tv_usec = 0;
 
 	if (run_mode) {
-		printf("port : %d\n", port);
+		printf("port : %lu\n", (unsigned long)port);
 	}
 
 	cs = pmap_rmtcall(&sin, progNum, VERSNUM, PROCNUM,
