+++ testcases/network/rpc/rpc-tirpc/tests_pack/rpc_suite/rpc/rpc_addrmanagmt_pmap_rmtcall/rpc_pmap_rmtcall_stress.c
@@ -76,7 +76,7 @@ int main(int argn, char *argc[])
 	tv.tv_usec = 100;
 
 	if (run_mode) {
-		printf("port : %d\n", port);
+		printf("port : %lu\n", (unsigned long)port);
 	}
 
 	for (i = 0; i < nbCall; i++) {
